feat(DeliverManagement): Introduce new components for order delivery management
- Add ErrorState, LoadingState, and EmptyState components to handle various delivery scenarios.
- Implement DeliverHeader for improved navigation and actions within the delivery management interface.
- Create OrderCard and OrderList components for displaying and managing orders with enhanced UI.
- Integrate TabsNavigation for organizing orders by status: processing, shipping, and delivered.
- Refactor existing code to utilize new utility functions for status normalization and date formatting.
- Enhance user experience with loading animations and error handling across components.
